With QueryByOrder tests
